Tell git to give more verbose output
When the external_project is cloning a git repo, git is detecting
that it's output stream is not a TTY, since it's captured by
execute_process, and therefore won't display updates to the progress of
clone operations as they happen; things like 1231/9878 objects
retrieved, etc.  If a repository is rather large, this can appear as
though something is hung since no progress can be seen being made.
Adding the --progress option to the git clone forces git to display the
progress output regardless of whether or not a TTY is detected.