SuperbuildInstallMacros: do not clean the prefix for install targets
Installing to /usr/local and nuking the prefix is a bad idea. The clean
bits are meant for packages which are contained safely in a new
directory.

Add a `--new` flag to start a new manifest and trigger `--clean` for
`CLEAN` based on whether the install target is being used or not.