hdf5: set HDF5_ROOT
This skips the config-based searching for HDF5 in FindHDF5. The config
module for HDF5 uses imported targets which causes issues in VTK's
module system if they are used (the third party support expects paths,
not targets), or on Windows, the paths to the `.dll` files end up on the
link line which doesn't work. Setting `HDF5_ROOT` ends up skipping all
of that and searching using `find_library` and such which works well
enough for the superbuild.