Make finding Python libs optional
In certain cases we want the Python interpreter to do other build
operations but we don't need the Python libraries. If a system
Python interpreter is available but not the Python libraries, without
this change we're foced to build Python in order to get both when we
really only want the interpreter.