windows: load manifest for dirs the executable will search for dlls
This adds an option to the windows fixup bundle script to specify the
directories that will be in the path when the executable is run.  The
CMake functions to install into the package default this to 'bin'.  This
allows the plugins to recognize that the ParaView libs they need are
already installed.