Developer-config: remove superfluous quotations from CMAKE_PREFIX_PATH
If CMAKE_PREFIX_PATH is set, then we set the xported CMAKE_PREFIX_PATH flag
to be a list of two things and it needs quotations. Otherwise, there is no
need to add quotations (and doing so results in a developer config file with
too many quotations, resulting in a warning).