SuperbuildPackageMacros: support package discovery and primacy
The CMB superbuild had "modes" based on which package was being built.
The superbuild now supports this as a first-class functionality by
storing "packages" in a directory and allowing selection between
available packages that are discovered. Each package may provide a
`<package>.configure.cmake` file which is included so that it may select
various options and such.